+++ .circleci/config.yml
@@ -3,7 +3,7 @@ version: 2.1
 jobs:
   build-client:
     docker:
-      - image: alexfalkowski/go:1.10
+      - image: alexfalkowski/go:1.12
       - image: bitnami/redis:7.4
         environment:
           REDIS_PASSWORD: test
@@ -61,7 +61,7 @@ jobs:
     resource_class: large
   build-docker:
     docker:
-      - image: alexfalkowski/go:1.10
+      - image: alexfalkowski/go:1.12
     working_directory: ~/servicectl
     steps:
       - checkout
@@ -84,7 +84,7 @@ jobs:
     resource_class: large
   push-docker:
     docker:
-      - image: alexfalkowski/go:1.10
+      - image: alexfalkowski/go:1.12
     working_directory: ~/servicectl
     steps:
       - checkout
