+++ Dockerfile
@@ -8,7 +8,7 @@ COPY go.mod go.sum ./
 RUN go mod download
 
 COPY . ./
-RUN CGO_ENABLED=0 go build -ldflags="-s -w -X 'github.com/alexfalkowski/servicectl/cmd.Version=${version}' -extldflags '-static'" -tags netgo -a -o servicectl main.go
+RUN CGO_ENABLED=0 go build -ldflags="-s -w -X 'github.com/alexfalkowski/servicectl/internal/cmd.Version=${version}' -extldflags '-static'" -tags netgo -a -o servicectl main.go
 
 FROM gcr.io/distroless/static
