+++ internal/cmd/crypto/ed25519/module.go
