+++ internal/cmd/crypto/hmac/module.go
