+++ internal/cmd/crypto/rsa/rsa.go
@@ -1,30 +1,30 @@
-package aes
+package rsa
 
 import (
 	"context"
 
-	"github.com/alexfalkowski/go-service/crypto/aes"
 	"github.com/alexfalkowski/go-service/crypto/rand"
+	"github.com/alexfalkowski/go-service/crypto/rsa"
 	"github.com/alexfalkowski/go-service/flags"
 	"github.com/alexfalkowski/go-service/meta"
 	"github.com/alexfalkowski/go-service/runtime"
-	"github.com/alexfalkowski/servicectl/cmd/os"
-	"github.com/alexfalkowski/servicectl/cmd/runner"
-	"github.com/alexfalkowski/servicectl/config"
+	"github.com/alexfalkowski/servicectl/internal/cmd/os"
+	"github.com/alexfalkowski/servicectl/internal/cmd/runner"
+	"github.com/alexfalkowski/servicectl/internal/config"
 	"go.uber.org/fx"
 	"go.uber.org/zap"
 )
 
 var (
-	// RotateFlag defines wether we should rotate the key or not.
+	// RotateFlag defines wether we should rotate the keys or not.
 	RotateFlag = flags.Bool()
 
-	// VerifyFlag defines wether we should verify the key or not.
+	// VerifyFlag defines wether we should verify the keys or not.
 	VerifyFlag = flags.Bool()
 )
 
 // Start for AES.
-func Start(lc fx.Lifecycle, logger *zap.Logger, rand *rand.Generator, gen *aes.Generator, cfg *config.Config) {
+func Start(lc fx.Lifecycle, logger *zap.Logger, rand *rand.Generator, gen *rsa.Generator, cfg *config.Config) {
 	var (
 		fn runner.StartFn
 		op string
@@ -33,18 +33,21 @@ func Start(lc fx.Lifecycle, logger *zap.Logger, rand *rand.Generator, gen *aes.G
 	switch {
 	case flags.IsBoolSet(RotateFlag):
 		fn = func(ctx context.Context) context.Context {
-			k, err := gen.Generate()
+			pub, pri, err := gen.Generate()
 			runtime.Must(err)
 
-			err = os.WriteBase64File(cfg.Crypto.AES.Key, []byte(k))
+			err = os.WriteFile(cfg.Crypto.RSA.Public, []byte(pub))
+			runtime.Must(err)
+
+			err = os.WriteFile(cfg.Crypto.RSA.Private, []byte(pri))
 			runtime.Must(err)
 
 			return ctx
 		}
-		op = "rotated key"
+		op = "rotated keys"
 	case flags.IsBoolSet(VerifyFlag):
 		fn = func(ctx context.Context) context.Context {
-			a, err := aes.NewCipher(rand, cfg.Crypto.AES)
+			a, err := rsa.NewCipher(rand, cfg.Crypto.RSA)
 			runtime.Must(err)
 
 			msg := "this is a test"
@@ -56,9 +59,9 @@ func Start(lc fx.Lifecycle, logger *zap.Logger, rand *rand.Generator, gen *aes.G
 
 			return meta.WithAttribute(ctx, "testMsg", meta.String(msg))
 		}
-		op = "verified key"
+		op = "verified keys"
 	}
 
 	opts := &runner.Options{Lifecycle: lc, Logger: logger, Fn: fn}
-	runner.Start("aes", op, opts)
+	runner.Start("rsa", op, opts)
 }
