+++ internal/cmd/hooks/hooks.go
@@ -10,9 +10,9 @@ import (
 	h "github.com/alexfalkowski/go-service/hooks"
 	"github.com/alexfalkowski/go-service/meta"
 	"github.com/alexfalkowski/go-service/runtime"
-	"github.com/alexfalkowski/servicectl/cmd/os"
-	"github.com/alexfalkowski/servicectl/cmd/runner"
-	"github.com/alexfalkowski/servicectl/config"
+	"github.com/alexfalkowski/servicectl/internal/cmd/os"
+	"github.com/alexfalkowski/servicectl/internal/cmd/runner"
+	"github.com/alexfalkowski/servicectl/internal/config"
 	hooks "github.com/standard-webhooks/standard-webhooks/libraries/go"
 	"go.uber.org/fx"
 	"go.uber.org/zap"
