+++ internal/cmd/hooks/module.go
